+++ ts/features/idpay/payment/components/Barcode/index.ts
@@ -1,15 +1,5 @@
-import { DecodedIOBarcode } from "./decoders";
 import { useIOBarcodeScanner } from "./useIOBarcodeScanner";
+import { IOBarcodeFormat, IOBarcodeType, IOBarcode } from "./types";
 
-export type IOBarcodeFormat = "DATA_MATRIX" | "QR_CODE";
-
-export type IOBarcodeType = DecodedIOBarcode["type"];
-
-/**
- * Scanned barcode, it contains the information about the scanned content, its format and its type
- */
-export type IOBarcode = {
-  format: IOBarcodeFormat;
-} & DecodedIOBarcode;
-
+export type { IOBarcodeType, IOBarcodeFormat, IOBarcode };
 export { useIOBarcodeScanner };
